Fix variables and closing braces in doinst
It is not possible to use the $pkgname and $pkgver variables inside the
doinst() function. These will be converted to their proper values before
writing the actual doinst.sh script.

Also, previously the doinst() function would read up to the first
closing brace that would be the only character in a line. But that brace
could easily not be the actual closing brace of the doinst() function.
Now the declare internal bash command is used to print the contents of
the doinst() function to the doinst.sh script.

This closes #1 and closes #2