Use declare to write the build() function
The build() function was extracted from the SLKBUILD using sed, reading
the line that starts with "build()" and ending at the first line that
has a closing brace as the first and only character in the line. But
this could very well not be the actual closing brace of the build()
function. So now, the declare internal bash command is used to write the
function to the buildscript, which fixes this.

This closes #4